\name{dataprocess}
\alias{dataprocess}
\title{return a data.frame}
\usage{
dataprocess (root,filename)
}
\arguments{
 \item{root}{the path of the file}
 \item{filename}{the name of QQ chating records}
}
\value{
return a data.frame,the fuction of timepoint(),speaker() and topic() use it.
}
\description{
return the a data.frame
}
\author{
heng wang
}
\examples{
#not run
#root<-"C:/Users/henry wang/Documents/"
#filename<-"records.txt"
#dataprocess (root,filename)
}
