\name{speaker}
\alias{speaker}
\title{
analysis the people what speake more.
}
\description{
speaker(QQdata)
}
\usage{
speaker(QQdata)
}
\arguments{
  \item{QQdata}{
  QQdata is the returned valua of dataprocess()
  }
}

\value{
plot a bar chart to describe  who speak more.
}
\author{
heng wang
}

\seealso{
 topic(), timepoint()
}
\examples{
#not run
#root<-"C:/Users/henry wang/Documents/"
#filename<-"records.txt"
#QQdata<-dataprocess (root,filename)
#speaker(QQdata)
}

