\name{timepoint}
\alias{timepoint}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  analytic time point
}
\description{
  timepoint(QQdata)
}
\usage{
timepoint(QQdata)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{QQdata}{
QQdata is the returned valua of dataprocess()
}
}

\value{
using ggplot2 plot a bar chart about time when people talked.
}
\references{
speaker(), topic()
}
\author{
heng wang
}

\examples{

#not run
#root<-"C:/Users/henry wang/Documents/"
#filename<-"records.txt"
#QQdata<-dataprocess (root,filename)
#timepoint(QQdata)
}

