\name{topic}
\alias{topic}
\title{
analiyse the topic.
}
\description{
topic()
}
\usage{
topic(QQdata)
}
\arguments{
  \item{QQdata}{
  QQdata is the reeturned valua of dataprocess()
  }
}

\value{
using the wordcloud2 to plot a wordcloud to describe  what they usully talk.
}
\author{
heng wang
}
\examples{
#not run
#root<-"C:/Users/henry wang/Documents/"
#filename<-"records.txt"
#QQdata<-dataprocess (root,filename)
#topic(QQdata)
}
