\name{dataprocess}
\alias{dataprocess}
<<<<<<< HEAD
\title{return a data.frame}
=======
\title{
dataprocess
}
\description{
return the a data.frame
}
>>>>>>> e7a21eaa37f7ed0413cd504a37ba35f8bde456f4
\usage{
dataprocess (root,filename)
}
\arguments{
 \item{root}{the path of the file}
 \item{filename}{the name of QQ chating records}
}
\value{
return a data.frame,the fuction of timepoint(),speaker() and topic() use it.
}
<<<<<<< HEAD
\description{
return the a data.frame
}
=======

>>>>>>> e7a21eaa37f7ed0413cd504a37ba35f8bde456f4
\author{
heng wang
}
\examples{
#not run
<<<<<<< HEAD
#root<-"C:/Users/henry wang/Documents/"
#filename<-"records.txt"
#dataprocess (root,filename)
=======
root<-"C:/Users/henry wang/Documents/"
filename<-"records.txt"
dataprocess (root,filename)
>>>>>>> e7a21eaa37f7ed0413cd504a37ba35f8bde456f4
}
